Fix calling convention alignment in web analysis
Consider the following program:

fun f(v1) = ...
... f(v10) ... f(v20) ...

There may be situations in which v20 does not have a binding, i.e., the
second call site is unreachable. The current web calculation would
ignore v20 and conclude that v1 and v10 belong to the same web and v20
belongs to its own web. This leads to inconsistent calling convention to
f if the v1-v10 web decides to flatten while v20 does not.